Slug extended to 16 chars; Format Update
