+++ packages/frontend/src/app/appTypes.ts
@@ -1,0 +1,22 @@
+import type { OrganizationRole } from "@festival/common";
+
+export interface InviteDraft {
+	email: string;
+	role: OrganizationRole;
+}
+
+export interface InviteFeedback {
+	id: number;
+	email: string;
+	role: OrganizationRole;
+	status: "success" | "error";
+}
+
+export interface FestivalDraft {
+	name: string;
+	startDate: string;
+	endDate: string;
+}
+
+export type SignInModalKind = "create-org" | "invite";
+export type SignInStep = "method" | "email";
